+++ astropy/wcs/docstrings.py
@@ -400,7 +400,8 @@ understood by the lower-level projection routines.
 """
 
 cunit = """
-``list of strings[naxis]`` List of ``CUNITia`` keyvalues.
+``list of astropy.UnitBase[naxis]`` List of ``CUNITia`` keyvalues as
+`astropy.units.UnitBase` instances.
 
 These define the units of measurement of the ``CRVALia``, ``CDELTia``
 and ``CDi_ja`` keywords.
@@ -2003,16 +2004,21 @@ a units specification where it does not create ambiguity (e.g. not
 between a metric prefix and a basic unit string), including in strings
 like ``"log (m ** 2)"`` which is formally disallowed.
 
+.. note:: Deprecated in Astropy 0.2
+
+   `UnitConverter` will be removed in a future version of astropy.
+   The `astropy.units` package should be used instead.
+
 Parameters
 ----------
 
 have : string
-    :ref:`fits-unit` to convert from, with or without surrounding
+    FITS unit string to convert from, with or without surrounding
     square brackets (for inline specifications); text following the
     closing bracket is ignored.
 
 want : string
-    :ref:`fits-unit` to convert to, with or without surrounding square
+    FITS unit string to convert to, with or without surrounding square
     brackets (for inline specifications); text following the closing
     bracket is ignored.
 
@@ -2036,8 +2042,6 @@ ctrl : string, optional
 
     Thus ``''`` doesn't do any unsafe translations, whereas ``'shd'``
     does all of them.
-
-    See :ref:`fits-unit` for more information.
 
 Raises
 ------
@@ -2108,8 +2112,6 @@ translate_units : string, optional
 
     Thus ``''`` doesn't do any unsafe translations, whereas ``'shd'``
     does all of them.
-
-    See :ref:`fits-unit` for more information.
 
 Returns
 -------