+++ astropy/wcs/setup_package.py
@@ -262,6 +262,7 @@ def get_extensions():
         'sip.c',
         'sip_wrap.c',
         'str_list_proxy.c',
+        'unit_list_proxy.c',
         'util.c',
         'wcslib_wrap.c',
         'wcslib_tabprm_wrap.c',